Migrate codebase JavaScript to TypeScript
