+++ package.json
@@ -2,6 +2,7 @@
   "name": "vuelve",
   "description": "A declarative syntax for the Composition API in Vue 3.",
   "version": "0.2.1",
+  "type": "module",
   "main": "dist/vuelve.cjs.js",
   "module": "dist/vuelve.esm.js",
   "types": "dist/vuelve.d.ts",
@@ -11,7 +12,7 @@
     "build": "tsc && vite build",
     "preview": "vite preview",
     "lint": "lint-staged",
-    "lint-all": "pretty-quick --pattern \"src/**/*.js\" --pattern \"test/**/*.js\" && eslint --fix src",
+    "lint-all": "pretty-quick --pattern \"src/**/*.{js,ts}\" --pattern \"test/**/*.{js,ts}\" && eslint --fix \"src/**/*.{js,ts}\"",
     "test": "jest --coverage",
     "test-watch": "jest --watch --coverage"
   },
@@ -36,20 +37,23 @@
     "lodash.clonedeep": "^4.5.0"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.12",
+    "@types/lodash.clonedeep": "^4.5.9",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
     "@vue/test-utils": "^2.4.5",
     "coveralls": "^3.1.0",
-    "eslint": "^7.12.1",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint": "^8.57.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^9.1.0",
     "husky": "^4.3.0",
-    "jest": "^26.6.1",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "lint-staged": "^10.5.0",
     "prettier": "^2.1.2",
     "pretty-quick": "^3.1.0",
+    "ts-jest": "^29.1.2",
+    "typescript": "^5.4.2",
     "vite": "^5.1.6",
     "vite-plugin-dts": "^3.7.3",
     "vue": "^3.0.2"
@@ -60,7 +64,7 @@
     }
   },
   "lint-staged": {
-    "src/*.js": [
+    "*.{js,jsx,ts,tsx}": [
       "pretty-quick --staged",
       "eslint --fix"
     ]
